308310 Prepare for very large arrays

+++ plugins/org.eclipse.mat.api/src/org/eclipse/mat/snapshot/inspections/MultiplePath2GCRootsQuery.java
@@ -207,7 +207,7 @@ public class MultiplePath2GCRootsQuery implements IQuery
         String attribute;
         String label;
         String gcRoots;
-        int shallowHeap;
+        long shallowHeap;
         long refShallowHeap;
         long retainedHeap;
